perf: Introduce memoized capped
