+++ apps/web/src/utils/memoizeCapped.ts
@@ -1,0 +1,14 @@
+import memoize from 'lodash/memoize'
+
+export const memoizeCapped = <T extends (...args: any) => any>(func: T) => {
+  const result = memoize(func, function (key) {
+    // @ts-ignore
+    if (cache?.size === 1000) {
+      cache.clear?.()
+    }
+    return key
+  })
+
+  const cache = result.cache
+  return result
+}
